# api/search/openapi.yaml
openapi: 3.0.3
info:
  title: NimoOS-Search API
  version: 0.1.0
  description: Retrieval API for NimoOS RAG layer. See design spec 2026-05-22-nimoos-search-design.md
servers:
  - url: http://localhost/v1/search
paths:
  /text:
    post:
      operationId: searchText
      summary: Text retrieval (BGE-M3 dense + sparse + rerank)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SearchResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  schemas:
    Filters:
      type: object
      properties:
        root_ids: { type: array, items: { type: string } }
        mime_prefix: { type: array, items: { type: string } }
        kind_in: { type: array, items: { type: string } }
        lang_in: { type: array, items: { type: string } }
    SearchRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        filters: { $ref: '#/components/schemas/Filters' }
        top_k: { type: integer, default: 20, maximum: 100 }
        rerank: { type: boolean, default: true }
        include_neighbors: { type: boolean, default: false }
        group_by_file: { type: boolean, default: false }
        max_chunks_per_file: { type: integer, default: 5, maximum: 20 }
    Cite:
      type: object
      properties:
        page: { type: integer, nullable: true }
        offset_start: { type: integer, nullable: true }
        offset_end: { type: integer, nullable: true }
        frame_ms_start: { type: integer, nullable: true }
        frame_ms_end: { type: integer, nullable: true }
        chunk_no: { type: integer }
    Hit:
      type: object
      properties:
        score: { type: number }
        raw_score: { type: number }
        collection: { type: string }
        file_id: { type: string }
        paths:
          type: array
          items:
            type: object
            properties:
              root_id: { type: string }
              path: { type: string }
        mime: { type: string }
        kind: { type: string }
        cite: { $ref: '#/components/schemas/Cite' }
        preview:
          type: object
          properties:
            text: { type: string, nullable: true }
            thumbnail_url: { type: string, nullable: true }
        payload_extra: { type: object }
    SearchResponse:
      type: object
      properties:
        hits:
          type: array
          items: { $ref: '#/components/schemas/Hit' }
        files:
          type: array
          items: { $ref: '#/components/schemas/FileGroup' }
        stats:
          type: object
          properties:
            total_candidates: { type: integer }
            rerank_ms: { type: integer }
            embed_ms: { type: integer }
            vector_search_ms: { type: integer }
            expand_ms: { type: integer }
        warnings:
          type: array
          items: { type: string }
...
    FileGroup:
      type: object
      properties:
        file_id: { type: string }
        paths:
          type: array
          items:
            type: object
            properties:
              root_id: { type: string }
              path: { type: string }
        mime: { type: string }
        kind: { type: string }
        score: { type: number }
        chunks:
          type: array
          items: { $ref: '#/components/schemas/Hit' }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    ServiceUnavailable:
      description: Upstream unavailable / not implemented
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              detail: { type: string, nullable: true }
              retry_after: { type: integer, nullable: true }